Initial version of treble support.
Defines the initial HIDL interface that will be used in place of the
old antradio.so bindings. An example server implementation is provided
as a starting point. Chip vendors are expected to implement and
maintain their own server implementations. The client of the interface
will be the new version of ANTHALService.